+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/implicits/RichCheckboxesSpec.scala
@@ -23,7 +23,9 @@ import uk.gov.hmrc.govukfrontend.views.viewmodels.checkboxes.{CheckboxItem, Chec
 import uk.gov.hmrc.govukfrontend.views.viewmodels.content.{HtmlContent, Text}
 import uk.gov.hmrc.govukfrontend.views.viewmodels.errormessage.ErrorMessage
 import uk.gov.hmrc.govukfrontend.views.viewmodels.errormessage.ErrorMessage.errorMessageWithDefaultStringsTranslated
+import uk.gov.hmrc.govukfrontend.views.viewmodels.fieldset.{Fieldset, Legend}
 import uk.gov.hmrc.helpers.MessagesHelpers
+import uk.gov.hmrc.hmrcfrontend.views.config.{HmrcPageHeadingLegend, HmrcSectionCaption}
 
 class RichCheckboxesSpec extends AnyWordSpec with Matchers with MessagesHelpers with RichFormInputHelpers {
 
@@ -110,6 +112,83 @@ class RichCheckboxesSpec extends AnyWordSpec with Matchers with MessagesHelpers 
         errorMessage = Some(ErrorMessage(content = Text("Checkboxes Error")))
       ).withFormFieldWithErrorAsHtml(field)
       checkboxes.errorMessage shouldBe Some(ErrorMessage(content = Text("Checkboxes Error")))
+    }
+  }
+
+  "Given a Checkboxes object, calling withHeading" should {
+    "set the fieldset legend to the passed in content" in {
+      val checkboxes = Checkboxes().withHeading(Text("This is a text heading"))
+      checkboxes shouldBe Checkboxes(fieldset =
+        Some(
+          Fieldset(
+            legend = Some(HmrcPageHeadingLegend(content = Text("This is a text heading")))
+          )
+        )
+      )
+    }
+
+    "set the fieldset legend to the passed in content, overriding any previously set content" in {
+      val checkboxesWithLegend =
+        Checkboxes(fieldset =
+          Some(Fieldset(legend = Some(Legend(content = Text("This is some original text heading")))))
+        )
+      val updatedCheckboxes    = checkboxesWithLegend.withHeading(Text("This is some updated text heading"))
+      updatedCheckboxes shouldBe Checkboxes(fieldset =
+        Some(
+          Fieldset(
+            legend = Some(HmrcPageHeadingLegend(content = Text("This is some updated text heading")))
+          )
+        )
+      )
+    }
+  }
+
+  "Given a Checkboxes object, calling withHeadingAndSectionCaption" should {
+    "set the fieldset legend to the passed in content" in {
+      val checkboxes =
+        Checkboxes().withHeadingAndSectionCaption(Text("This is a text heading"), Text("This is a text caption"))
+      checkboxes shouldBe Checkboxes(fieldset =
+        Some(
+          Fieldset(
+            legend = Some(
+              HmrcPageHeadingLegend(
+                content = Text("This is a text heading"),
+                caption = HmrcSectionCaption(Text("This is a text caption"))
+              )
+            )
+          )
+        )
+      )
+    }
+
+    "set the fieldset legend to the passed in content, overriding any previously set content" in {
+      val checkboxesWithLegend = Checkboxes(fieldset =
+        Some(
+          Fieldset(legend =
+            Some(
+              Legend(
+                content = Text("This is some original text heading")
+              )
+            )
+          )
+        )
+      )
+      val updatedCheckboxes    = checkboxesWithLegend.withHeadingAndSectionCaption(
+        Text("This is some updated text heading"),
+        Text("This is some updated text caption")
+      )
+      updatedCheckboxes shouldBe Checkboxes(fieldset =
+        Some(
+          Fieldset(
+            legend = Some(
+              HmrcPageHeadingLegend(
+                content = Text("This is some updated text heading"),
+                caption = HmrcSectionCaption(Text("This is some updated text caption"))
+              )
+            )
+          )
+        )
+      )
     }
   }
 }