+++ src/test/scala/uk/gov/hmrc/govukfrontend/views/implicits/RichTextareaSpec.scala
@@ -22,8 +22,10 @@ import uk.gov.hmrc.govukfrontend.views.html.components.implicits._
 import uk.gov.hmrc.govukfrontend.views.viewmodels.content.{HtmlContent, Text}
 import uk.gov.hmrc.govukfrontend.views.viewmodels.errormessage.ErrorMessage
 import uk.gov.hmrc.govukfrontend.views.viewmodels.errormessage.ErrorMessage.errorMessageWithDefaultStringsTranslated
+import uk.gov.hmrc.govukfrontend.views.viewmodels.label.Label
 import uk.gov.hmrc.govukfrontend.views.viewmodels.textarea.Textarea
 import uk.gov.hmrc.helpers.MessagesHelpers
+import uk.gov.hmrc.hmrcfrontend.views.config.{HmrcPageHeadingLabel, HmrcSectionCaption}
 
 class RichTextareaSpec extends AnyWordSpec with Matchers with MessagesHelpers with RichFormInputHelpers {
 
@@ -97,6 +99,57 @@ class RichTextareaSpec extends AnyWordSpec with Matchers with MessagesHelpers wi
         errorMessage = Some(ErrorMessage(content = Text("Textarea Error")))
       ).withFormFieldWithErrorAsHtml(field)
       textarea.errorMessage shouldBe Some(ErrorMessage(content = Text("Textarea Error")))
+    }
+  }
+
+  "Given a Textarea object, calling withHeading" should {
+    "set the label to the passed in content" in {
+      val textarea = Textarea().withHeading(heading = Text("This is a text heading"))
+      textarea shouldBe Textarea(
+        label = HmrcPageHeadingLabel(content = Text("This is a text heading"))
+      )
+    }
+
+    "set the label to the passed in content, overriding any previously set content" in {
+      val textareaWithLabel = Textarea(
+        label = Label(content = Text("This is some original text heading"))
+      )
+      val updatedTextarea   =
+        textareaWithLabel.withHeading(heading = Text("This is some updated text heading"))
+      updatedTextarea shouldBe Textarea(
+        label = HmrcPageHeadingLabel(content = Text("This is some updated text heading"))
+      )
+    }
+  }
+
+  "Given a Textarea object, calling withHeadingAndSectionCaption" should {
+    "set the label and caption to the passed in content" in {
+      val textarea = Textarea().withHeadingAndSectionCaption(
+        heading = Text("This is a text heading"),
+        sectionCaption = Text("This is a text caption")
+      )
+      textarea shouldBe Textarea(
+        label = HmrcPageHeadingLabel(
+          content = Text("This is a text heading"),
+          caption = HmrcSectionCaption(Text("This is a text caption"))
+        )
+      )
+    }
+
+    "set the label to the passed in content, overriding any previously set content" in {
+      val textareaWithLabel = Textarea(
+        label = Label(content = Text("This is some original text heading"))
+      )
+      val updatedTextarea   = textareaWithLabel.withHeadingAndSectionCaption(
+        heading = Text("This is some updated text heading"),
+        sectionCaption = Text("This is a text caption")
+      )
+      updatedTextarea shouldBe Textarea(
+        label = HmrcPageHeadingLabel(
+          content = Text("This is some updated text heading"),
+          caption = HmrcSectionCaption(Text("This is a text caption"))
+        )
+      )
     }
   }
 }